refactor: consolidate results panel onto shared grid modules
The panel kept private copies of helpers that gridCells.ts/gridTheme.ts
already export (forked while the notebook renderer was being added).
Import the shared cssVar/buildGridTheme/compareRows/type helpers and route
getCellContent through buildDisplayCell; move getColumnType into gridCells.
~110 fewer lines, no behavior change to the grid surface.
