+++ media/notebook-renderer.js
@@ -188,7 +188,7 @@ Please report this to https://github.com/markedjs/marked.`,e){let i="<p>An error
 `),rawValue:Q});return ze===void 0?void 0:(dn(ze.kind===z.kind),{location:F,value:ze})}}}let E=U.columns,B=U.rows,j=c.current?.contains(document.activeElement)===!0||a.current?.contains(document.activeElement)===!0,G;if(U.current!==void 0?G=[U.current.range.x,U.current.range.y]:E.length===1?G=[E.first()??0,0]:B.length===1&&(G=[H,B.first()??0]),j&&G!==void 0){let z,F,Q="text/plain",ie="text/html";if(navigator.clipboard.read!==void 0){let ze=await navigator.clipboard.read();for(let nt of ze){if(nt.types.includes(ie)){let on=await(await nt.getType(ie)).text(),It=qg(on);if(It!==void 0){z=It;break}}nt.types.includes(Q)&&(F=await(await nt.getType(Q)).text())}}else if(navigator.clipboard.readText!==void 0)F=await navigator.clipboard.readText();else if(w!==void 0&&w?.clipboardData!==null){if(w.clipboardData.types.includes(ie)){let ze=w.clipboardData.getData(ie);z=qg(ze)}z===void 0&&w.clipboardData.types.includes(Q)&&(F=w.clipboardData.getData(Q))}else return;let[Se,We]=G,Oe=[];do{if(xe===void 0){let ze=ur(G),nt=F??z?.map(on=>on.map(It=>It.rawValue).join("	")).join("	")??"",gt=T(ze,G,nt,void 0);gt!==void 0&&Oe.push(gt);break}if(z===void 0){if(F===void 0)return;z=Qb(F)}if(xe===!1||typeof xe=="function"&&xe?.([G[0]-H,G[1]],z.map(ze=>ze.map(nt=>nt.rawValue?.toString()??"")))!==!0)return;for(let[ze,nt]of z.entries()){if(ze+We>=v)break;for(let[gt,on]of nt.entries()){let It=[gt+Se,ze+We],[Nl,Ja]=It;if(Nl>=Be.length||Ja>=xn)continue;let es=ur(It),ri=T(es,It,on.rawValue,on.formatted);ri!==void 0&&Oe.push(ri)}}}while(!1);wr(Oe),jn.current?.damage(Oe.map(ze=>({cell:ze.location})))}},[M,Re,ur,U,zr.paste,Be.length,wr,xn,xe,H,v]);nn("paste",Jf,f,!1,!0);let Za=$.useCallback(async(w,T)=>{if(!zr.copy)return;let E=T===!0||c.current?.contains(document.activeElement)===!0||a.current?.contains(document.activeElement)===!0,B=U.columns,j=U.rows,G=(z,F)=>{if(!ut)$g(z,F,w);else{let Q=F.map(ie=>({kind:te.Text,data:m[ie].title,displayData:m[ie].title,allowOverlay:!1}));$g([Q,...z],F,w)}};if(E&&Jt!==void 0){if(U.current!==void 0){let z=Jt(U.current.range,qn.current.signal);typeof z!="object"&&(z=await z()),G(z,(0,hl.default)(U.current.range.x-H,U.current.range.x+U.current.range.width-H))}else if(j!==void 0&&j.length>0){let F=[...j].map(Q=>{let ie=Jt({x:H,y:Q,width:m.length,height:1},qn.current.signal);return typeof ie=="object"?ie[0]:ie().then(Se=>Se[0])});if(F.some(Q=>Q instanceof Promise)){let Q=await Promise.all(F);G(Q,(0,hl.default)(m.length))}else G(F,(0,hl.default)(m.length))}else if(B.length>0){let z=[],F=[];for(let Q of B){let ie=Jt({x:Q,y:0,width:1,height:v},qn.current.signal);typeof ie!="object"&&(ie=await ie()),z.push(ie),F.push(Q-H)}if(z.length===1)G(z[0],F);else{let Q=z.reduce((ie,Se)=>ie.map((We,Oe)=>[...We,...Se[Oe]]));G(Q,F)}}}},[m,Jt,U,zr.copy,H,v,ut]);nn("copy",Za,f,!1,!1);let U3=$.useCallback(async w=>{if(!(!zr.cut||!(c.current?.contains(document.activeElement)===!0||a.current?.contains(document.activeElement)===!0))&&(await Za(w),U.current!==void 0)){let E={current:{cell:U.current.cell,range:U.current.range,rangeStack:[]},rows:rt.empty(),columns:rt.empty()},B=P?.(E);if(B===!1||(E=B===!0?E:B,E.current===void 0))return;Po(E.current.range)}},[Po,U,zr.cut,Za,P]);nn("cut",U3,f,!1,!1);let G3=$.useCallback((w,T)=>{if(le!==void 0){H!==0&&(w=w.map(j=>[j[0]-H,j[1]])),le(w,T);return}if(w.length===0||T===-1)return;let[E,B]=w[T];d.current!==void 0&&d.current[0]===E&&d.current[1]===B||(d.current=[E,B],ji(E,B,!1,!1))},[le,H,ji]),[Vu,Uu]=Qt?.current?.cell??[],Gu=$.useRef(Gt);Gu.current=Gt,$.useLayoutEffect(()=>{!ct.current&&Vu!==void 0&&Uu!==void 0&&(Vu!==qi.current?.current?.cell[0]||Uu!==qi.current?.current?.cell[1])&&Gu.current(Vu,Uu),ct.current=!1},[Vu,Uu]);let y0=U.current!==void 0&&(U.current.cell[0]>=Be.length||U.current.cell[1]>=xn);$.useLayoutEffect(()=>{y0&&$n(If,!1)},[y0,$n]);let K3=$.useMemo(()=>Ot===!0&&it?.tint===!0?rt.fromSingleSelection(xn-1):rt.empty(),[xn,Ot,it?.tint]),X3=$.useCallback(w=>typeof Tt=="boolean"?Tt:Tt?.(w-H)??!0,[H,Tt]),Y3=$.useMemo(()=>{if(Gf===void 0||a.current===null)return null;let{bounds:w,group:T}=Gf,E=a.current.getBoundingClientRect();return $.createElement(qb,{bounds:w,group:T,canvasBounds:E,onClose:()=>Kf(void 0),onFinish:B=>{Kf(void 0),K?.(T,B)}})},[K,Gf]),Q3=Math.min(Be.length,Ve+(ar?1:0));$.useImperativeHandle(t,()=>({appendRow:(w,T)=>Dl(w+H,T),updateCells:w=>(H!==0&&(w=w.map(T=>({cell:[T.cell[0]+H,T.cell[1]]}))),jn.current?.damage(w)),getBounds:(w,T)=>{if(!(a?.current===null||c?.current===null)){if(w===void 0&&T===void 0){let E=a.current.getBoundingClientRect(),B=E.width/c.current.clientWidth;return{x:E.x-c.current.scrollLeft*B,y:E.y-c.current.scrollTop*B,width:c.current.scrollWidth*B,height:c.current.scrollHeight*B}}return jn.current?.getBounds((w??0)+H,T)}},focus:()=>jn.current?.focus(),emit:async w=>{switch(w){case"delete":Qa({bounds:void 0,cancel:()=>{},stopPropagation:()=>{},preventDefault:()=>{},ctrlKey:!1,key:"Delete",keyCode:46,metaKey:!1,shiftKey:!1,altKey:!1,rawEvent:void 0,location:void 0});break;case"fill-right":Qa({bounds:void 0,cancel:()=>{},stopPropagation:()=>{},preventDefault:()=>{},ctrlKey:!0,key:"r",keyCode:82,metaKey:!1,shiftKey:!1,altKey:!1,rawEvent:void 0,location:void 0});break;case"fill-down":Qa({bounds:void 0,cancel:()=>{},stopPropagation:()=>{},preventDefault:()=>{},ctrlKey:!0,key:"d",keyCode:68,metaKey:!1,shiftKey:!1,altKey:!1,rawEvent:void 0,location:void 0});break;case"copy":await Za(void 0,!0);break;case"paste":await Jf();break}},scrollTo:Gt,remeasureColumns:w=>{for(let T of w)ju(T+H)}}),[Dl,ju,Za,Qa,Jf,H,Gt]);let[w0,x0]=Ll??[],Z3=$.useCallback(w=>{let[T,E]=w;if(E===-1){q!=="none"&&(se(rt.fromSingleSelection(T),void 0,!1),Dn());return}w0===T&&x0===E||(Y({cell:w,range:{x:T,y:E,width:1,height:1}},!0,!1,"keyboard-nav"),Gt(T,E))},[q,Dn,Gt,w0,x0,Y,se]),[J3,eT]=$.useState(!1),S0=$.useRef((0,mE.default)(w=>{eT(w)},5)),tT=$.useCallback(()=>{S0.current(!0),U.current===void 0&&U.columns.length===0&&U.rows.length===0&&s===void 0&&Y({cell:[H,Fu],range:{x:H,y:Fu,width:1,height:1}},!0,!1,"keyboard-select")},[Fu,U,s,H,Y]),nT=$.useCallback(()=>{S0.current(!1)},[]),[rT,iT]=$.useMemo(()=>{let w,T=Wn?.scrollbarWidthOverride??uu(),E=v+(Ot?1:0);if(typeof Pn=="number")w=ht+E*Pn;else{let j=0,G=Math.min(E,10);for(let z=0;z<G;z++)j+=Pn(z);j=Math.floor(j/G),w=ht+E*j}w+=T;let B=Be.reduce((j,G)=>G.width+j,0)+T;return[`${Math.min(1e5,B)}px`,`${Math.min(1e5,w)}px`]},[Be,Wn?.scrollbarWidthOverride,Pn,v,Ot,ht]),oT=$.useMemo(()=>cf(Z),[Z]);return $.createElement(ff.Provider,{value:Z},$.createElement(eE,{style:oT,className:de,inWidth:C??rT,inHeight:g??iT},$.createElement(Wb,{fillHandle:Zt,drawFocusRing:rn,experimental:Wn,fixedShadowX:Mu,fixedShadowY:wl,getRowThemeOverride:p,headerIcons:xi,imageWindowLoader:Si,initialSize:xl,isDraggable:Iu,onDragLeave:_u,onRowMoved:Jr,overscrollX:Ff,overscrollY:Lu,preventDiagonalScrolling:Cl,rightElement:Ro,rightElementProps:Hi,smoothScrollX:Du,smoothScrollY:zf,className:de,enableGroups:Ft,onCanvasFocused:tT,onCanvasBlur:nT,canvasRef:a,onContextMenu:V3,theme:Z,cellXOffset:Mo,cellYOffset:Fu,accessibilityHeight:sn.height,onDragEnd:W3,columns:Be,nonGrowWidth:$t,drawHeader:qf,onColumnProposeMove:A,drawCell:$f,disabledRows:K3,freezeColumns:Q3,lockColumns:H,firstColAccessible:H===0,getCellContent:ur,minColumnWidth:lr,maxColumnWidth:Or,searchInputRef:l,showSearch:Va,onSearchClose:Bf,highlightRegions:D3,getCellsForSelection:Jt,getGroupDetails:Vf,headerHeight:Tl,isFocused:J3,groupHeaderHeight:Ft?ja:0,freezeTrailingRows:Nt+(Ot&&it?.sticky===!0?1:0),hasAppendRow:Ot,onColumnResize:Ri,onColumnResizeEnd:Wf,onColumnResizeStart:Il,onCellFocused:Z3,onColumnMoved:H3,onDragStart:B3,onHeaderMenuClick:N3,onItemHovered:Zf,isFilling:s?.fillHandle===!0,onMouseMove:L3,onKeyDown:Qa,onKeyUp:Ze,onMouseDown:O3,onMouseUp:A3,onDragOverCell:wt,onDrop:Zr,onSearchResultsChanged:G3,onVisibleRegionChanged:F3,clientSize:he,rowHeight:Pn,searchResults:Le,searchValue:fe,onSearchValueChange:ve,rows:xn,scrollRef:c,selection:U,translateX:sn.tx,translateY:sn.ty,verticalBorder:X3,gridRef:jn,getCellRenderer:Re}),Y3,i!==void 0&&$.createElement($.Suspense,{fallback:null},$.createElement(vL,{...i,validateCell:Ua,bloom:O,id:j3,getCellRenderer:Re,className:Wn?.isSubGrid===!0?"click-outside-ignore":void 0,provideEditor:Et,imageEditorOverride:h,onFinishEditing:$3,markdownDivCreateNode:y,isOutsideClick:Ou}))))},gE=$.forwardRef(xL);kt();var vE=20,yE={getAccessibilityString:e=>e.data?.toString()??"false",kind:te.Boolean,needsHover:!0,useLabel:!1,needsHoverPosition:!0,measure:()=>50,draw:e=>SL(e,e.cell.data,nu(e.cell),e.cell.maxSize??vE),onDelete:e=>({...e,data:!1}),onClick:e=>{let{cell:t,posX:n,posY:r,bounds:i,theme:o}=e,{width:l,height:a,x:s,y:u}=i,c=t.maxSize??vE,d=Math.floor(i.y+a/2),f=Yd(c,a,o.cellVerticalPadding),h=Xd(t.contentAlign??"center",s,l,o.cellHorizontalPadding,f),p=Kd(h,d,f),y=Qd(s+n,u+r,p);if(nu(t)&&y)return{...t,data:Tf(t.data)}},onPaste:(e,t)=>{let n=ka;return e.toLowerCase()==="true"?n=!0:e.toLowerCase()==="false"?n=!1:e.toLowerCase()==="indeterminate"&&(n=Ca),n===t.data?void 0:{...t,data:n}}};function SL(e,t,n,r){if(!n&&t===ka)return;let{ctx:i,hoverAmount:o,theme:l,rect:a,highlighted:s,hoverX:u,hoverY:c,cell:{contentAlign:d}}=e,{x:f,y:h,width:p,height:y}=a,C=.35,g=n?1-C+C*o:.4;t===ka&&(g*=o),g!==0&&(i.globalAlpha=g,Ia(i,l,t,f,h,p,y,s,u,c,r,d),i.globalAlpha=1)}var SE=ke(Ie(),1);var _f=ke(Ie(),1);Yn();var wE=vt("div")({name:"BubblesOverlayEditorStyle",class:"gdg-b1ygi5by",propsAsIs:!1});var kL=e=>{let{bubbles:t}=e;return _f.createElement(wE,null,t.map((n,r)=>_f.createElement("div",{key:r,className:"boe-bubble"},n)),_f.createElement("textarea",{className:"gdg-input",autoFocus:!0}))},xE=kL;kt();var kE={getAccessibilityString:e=>Zd(e.data),kind:te.Bubble,needsHover:!1,useLabel:!1,needsHoverPosition:!1,measure:(e,t,n)=>t.data.reduce((r,i)=>e.measureText(i).width+r+20,0)+2*n.cellHorizontalPadding-4,draw:e=>RL(e,e.cell.data),provideEditor:()=>e=>{let{value:t}=e;return SE.createElement(xE,{bubbles:t.data})},onPaste:()=>{}},CL=4;function RL(e,t){let{rect:n,theme:r,ctx:i,highlighted:o}=e,{x:l,y:a,width:s,height:u}=n,c=20,d=8,f=CL,h=l+r.cellHorizontalPadding,p=[];for(let y of t){if(h>l+s)break;let C=Fi(y,i,r.baseFontFull).width;p.push({x:h,width:C}),h+=C+d*2+f}i.beginPath();for(let y of p)In(i,y.x,a+(u-c)/2,y.width+d*2,c,r.roundingRadius??c/2);i.fillStyle=o?r.bgBubbleSelected:r.bgBubble,i.fill();for(let[y,C]of p.entries())i.beginPath(),i.fillStyle=r.textBubble,i.fillText(t[y],C.x+d,a+u/2+Zn(i,r))}var RE=ke(Ie(),1);var Eu=ke(Ie(),1);Yn();var bL=vt("div")({name:"DrilldownOverlayEditorStyle",class:"gdg-d4zsq0x",propsAsIs:!1}),EL=e=>{let{drilldowns:t}=e;return Eu.createElement(bL,null,t.map((n,r)=>Eu.createElement("div",{key:r,className:"doe-bubble"},n.img!==void 0&&Eu.createElement("img",{src:n.img}),Eu.createElement("div",null,n.text))))},CE=EL;kt();var bE={getAccessibilityString:e=>Zd(e.data.map(t=>t.text)),kind:te.Drilldown,needsHover:!1,useLabel:!1,needsHoverPosition:!1,measure:(e,t,n)=>t.data.reduce((r,i)=>e.measureText(i.text).width+r+20+(i.img!==void 0?18:0),0)+2*n.cellHorizontalPadding-4,draw:e=>IL(e,e.cell.data),provideEditor:()=>e=>{let{value:t}=e;return RE.createElement(CE,{drilldowns:t.data})},onPaste:()=>{}},TL=4,Gg={};function ML(e,t,n,r){let i=Math.ceil(window.devicePixelRatio),o=5,l=n-o*2,a=4,s=n*i,u=r+o,c=r*3,d=(c+o*2)*i,f=`${e},${t},${i},${n}`;if(Gg[f]!==void 0)return{el:Gg[f],height:s,width:d,middleWidth:a*i,sideWidth:u*i,padding:o*i,dpr:i};let h=document.createElement("canvas"),p=h.getContext("2d");return p===null?null:(h.width=d,h.height=s,p.scale(i,i),Gg[f]=h,p.beginPath(),In(p,o,o,c,l,r),p.shadowColor="rgba(24, 25, 34, 0.4)",p.shadowBlur=1,p.fillStyle=e,p.fill(),p.shadowColor="rgba(24, 25, 34, 0.3)",p.shadowOffsetY=1,p.shadowBlur=5,p.fillStyle=e,p.fill(),p.shadowOffsetY=0,p.shadowBlur=0,p.shadowBlur=0,p.beginPath(),In(p,o+.5,o+.5,c,l,r),p.strokeStyle=t,p.lineWidth=1,p.stroke(),{el:h,height:s,width:d,sideWidth:u*i,middleWidth:r*i,padding:o*i,dpr:i})}function IL(e,t){let{rect:n,theme:r,ctx:i,imageLoader:o,col:l,row:a}=e,{x:s,width:u}=n,c=r.baseFontFull,d=Xm(i,c),f=Math.min(n.height,Math.max(16,Math.ceil(d*r.lineHeight)*2)),h=Math.floor(n.y+(n.height-f)/2),p=f-10,y=8,C=TL,g=s+r.cellHorizontalPadding,m=r.roundingRadius??6,v=ML(r.bgCell,r.drilldownBorder,f,m),x=[];for(let S of t){if(g>s+u)break;let k=Fi(S.text,i,c).width,b=0;S.img!==void 0&&o.loadOrGetImage(S.img,l,a)!==void 0&&(b=p-8+4);let M=k+b+y*2;x.push({x:g,width:M}),g+=M+C}if(v!==null){let{el:S,height:R,middleWidth:k,sideWidth:b,width:M,dpr:_,padding:L}=v,O=b/_,D=L/_;for(let A of x){let ne=Math.floor(A.x),Ce=Math.floor(A.width),ce=Ce-(O-D)*2;i.imageSmoothingEnabled=!1,i.drawImage(S,0,0,b,R,ne-D,h,O,f),ce>0&&i.drawImage(S,b,0,k,R,ne+(O-D),h,ce,f),i.drawImage(S,M-b,0,b,R,ne+Ce-(O-D),h,O,f),i.imageSmoothingEnabled=!0}}i.beginPath();for(let[S,R]of x.entries()){let k=t[S],b=R.x+y;if(k.img!==void 0){let M=o.loadOrGetImage(k.img,l,a);if(M!==void 0){let _=p-8,L=0,O=0,D=M.width,A=M.height;D>A?(L+=(D-A)/2,D=A):A>D&&(O+=(A-D)/2,A=D),i.beginPath(),In(i,b,h+f/2-_/2,_,_,r.roundingRadius??3),i.save(),i.clip(),i.drawImage(M,L,O,D,A,b,h+f/2-_/2,_,_),i.restore(),b+=_+4}}i.beginPath(),i.fillStyle=r.textBubble,i.fillText(k.text,b,h+f/2+Zn(i,r))}}var EE=ke(Ie(),1);kt();var TE={getAccessibilityString:e=>e.data.join(", "),kind:te.Image,needsHover:!1,useLabel:!1,needsHoverPosition:!1,draw:e=>_L(e,e.cell.displayData??e.cell.data,e.cell.rounding??e.theme.roundingRadius??4,e.cell.contentAlign),measure:(e,t)=>t.data.length*50,onDelete:e=>({...e,data:[]}),provideEditor:()=>e=>{let{value:t,onFinishedEditing:n,imageEditorOverride:r}=e;return EE.createElement(r??QS,{urls:t.data,canWrite:t.readonly!==!1,onCancel:n,onChange:o=>{n({...t,data:[o]})}})},onPaste:(e,t)=>{e=e.trim();let r=e.split(",").map(i=>{try{return new URL(i),i}catch{return}}).filter(i=>i!==void 0);if(!(r.length===t.data.length&&r.every((i,o)=>i===t.data[o])))return{...t,data:r}}},Kg=4;function _L(e,t,n,r){let{rect:i,col:o,row:l,theme:a,ctx:s,imageLoader:u}=e,{x:c,y:d,height:f,width:h}=i,p=f-a.cellVerticalPadding*2,y=[],C=0;for(let m=0;m<t.length;m++){let v=t[m];if(v.length===0)continue;let x=u.loadOrGetImage(v,o,l);if(x!==void 0){y[m]=x;let S=x.width*(p/x.height);C+=S+Kg}}if(C===0)return;C-=Kg;let g=c+a.cellHorizontalPadding;r==="right"?g=Math.floor(c+h-a.cellHorizontalPadding-C):r==="center"&&(g=Math.floor(c+h/2-C/2));for(let m of y){if(m===void 0)continue;let v=m.width*(p/m.height);n>0&&(s.beginPath(),In(s,g,d+a.cellVerticalPadding,v,p,n),s.save(),s.clip()),s.drawImage(m,g,d+a.cellVerticalPadding,v,p),n>0&&s.restore(),g+=v+Kg}}fi();kt();function PL(e,t){let n=e*49632+t*325176;return n^=n<<13,n^=n>>17,n^=n<<5,n/4294967295*2}var ME={getAccessibilityString:()=>"",kind:te.Loading,needsHover:!1,useLabel:!1,needsHoverPosition:!1,measure:()=>120,draw:e=>{let{cell:t,col:n,row:r,ctx:i,rect:o,theme:l}=e;if(t.skeletonWidth===void 0||t.skeletonWidth===0)return;let a=t.skeletonWidth;t.skeletonWidthVariability!==void 0&&t.skeletonWidthVariability>0&&(a+=Math.round(PL(n,r)*t.skeletonWidthVariability));let s=l.cellHorizontalPadding,u=t.skeletonHeight??Math.min(18,o.height-2*l.cellVerticalPadding);In(i,o.x+s,o.y+(o.height-u)/2,a,u,l.roundingRadius??3),i.fillStyle=Ir(l.textDark,.1),i.fill()},onPaste:()=>{}};var _E=ke(Ie(),1);var fn=ke(Ie(),1);Yn();var DL=()=>e=>e.targetWidth,Xg=vt("div")({name:"MarkdownOverlayEditorStyle",class:"gdg-m1pnx84e",propsAsIs:!1,vars:{"m1pnx84e-0":[DL(),"px"]}});var IE=e=>{let{value:t,onChange:n,forceEditMode:r,createNode:i,targetRect:o,onFinish:l,validatedSelection:a}=e,s=t.data,u=t.readonly===!0,[c,d]=fn.useState(s===""||r),f=fn.useCallback(()=>{d(p=>!p)},[]),h=s?"gdg-ml-6":"";return c?fn.createElement(Xg,{targetWidth:o.width-20},fn.createElement(go,{autoFocus:!0,highlight:!1,validatedSelection:a,value:s,onKeyDown:p=>{p.key==="Enter"&&p.stopPropagation()},onChange:n}),fn.createElement("div",{className:`gdg-edit-icon gdg-checkmark-hover ${h}`,onClick:()=>l(t)},fn.createElement(US,null))):fn.createElement(Xg,{targetWidth:o.width},fn.createElement(hu,{contents:s,createNode:i}),!u&&fn.createElement(fn.Fragment,null,fn.createElement("div",{className:"spacer"}),fn.createElement("div",{className:`gdg-edit-icon gdg-edit-hover ${h}`,onClick:f},fn.createElement(Ra,null))),fn.createElement("textarea",{className:"gdg-md-edit-textarea gdg-input",autoFocus:!0}))};kt();var PE={getAccessibilityString:e=>e.data?.toString()??"",kind:te.Markdown,needsHover:!1,needsHoverPosition:!1,drawPrep:pi,measure:(e,t,n)=>{let r=t.data.split(`
 `)[0];return e.measureText(r).width+2*n.cellHorizontalPadding},draw:e=>Jn(e,e.cell.data,e.cell.contentAlign),onDelete:e=>({...e,data:""}),provideEditor:()=>e=>{let{onChange:t,value:n,target:r,onFinishedEditing:i,markdownDivCreateNode:o,forceEditMode:l,validatedSelection:a}=e;return _E.createElement(IE,{onFinish:i,targetRect:r,value:n,validatedSelection:a,onChange:s=>t({...n,data:s.target.value}),forceEditMode:l,createNode:o})},onPaste:(e,t)=>e===t.data?void 0:{...t,data:e}};kt();var DE={getAccessibilityString:e=>e.row.toString(),kind:En.Marker,needsHover:!0,needsHoverPosition:!1,drawPrep:OL,measure:()=>44,draw:e=>AL(e,e.cell.row,e.cell.checked,e.cell.markerKind,e.cell.drawHandle,e.cell.checkboxStyle),onClick:e=>{let{bounds:t,cell:n,posX:r,posY:i}=e,{width:o,height:l}=t,a=n.drawHandle?7+(o-7)/2:o/2,s=l/2;if(Math.abs(r-a)<=10&&Math.abs(i-s)<=10)return{...n,checked:!n.checked}},onPaste:()=>{}};function OL(e,t){let{ctx:n,theme:r}=e,i=r.markerFontFull,o=t??{};return o?.font!==i&&(n.font=i,o.font=i),o.deprep=zL,n.textAlign="center",o}function zL(e){let{ctx:t}=e;t.textAlign="start"}function AL(e,t,n,r,i,o){let{ctx:l,rect:a,hoverAmount:s,theme:u}=e,{x:c,y:d,width:f,height:h}=a,p=n?1:r==="checkbox-visible"?.6+.4*s:s;if(r!=="number"&&p>0){l.globalAlpha=p;let y=7*(n?s:1);if(Ia(l,u,n,i?c+y:c,d,i?f-y:f,h,!0,void 0,void 0,18,"center",o),i){l.globalAlpha=s,l.beginPath();for(let C of[3,6])for(let g of[-5,-1,3])l.rect(c+C,d+h/2+g,2,2);l.fillStyle=u.textLight,l.fill(),l.beginPath()}l.globalAlpha=1}if(r==="number"||r==="both"&&!n){let y=t.toString(),C=u.markerFontFull,g=c+f/2;r==="both"&&s!==0&&(l.globalAlpha=1-s),l.fillStyle=u.textLight,l.font=C,l.fillText(y,g,d+h/2+Zn(l,C)),s!==0&&(l.globalAlpha=1)}}kt();var OE={getAccessibilityString:()=>"",kind:En.NewRow,needsHover:!0,needsHoverPosition:!1,measure:()=>200,draw:e=>LL(e,e.cell.hint,e.cell.icon),onPaste:()=>{}};function LL(e,t,n){let{ctx:r,rect:i,hoverAmount:o,theme:l,spriteManager:a}=e,{x:s,y:u,width:c,height:d}=i;r.beginPath(),r.globalAlpha=o,r.rect(s+1,u+1,c,d-2),r.fillStyle=l.bgHeaderHovered,r.fill(),r.globalAlpha=1,r.beginPath();let f=t!=="",h=0;if(n!==void 0){let y=d-8,C=s+8/2,g=u+8/2;a.drawSprite(n,"normal",r,C,g,y,l,f?1:o),h=y}else{h=24;let p=12,y=f?p:o*p,C=f?0:(1-o)*p*.5,g=l.cellHorizontalPadding+4;y>0&&(r.moveTo(s+g+C,u+d/2),r.lineTo(s+g+C+y,u+d/2),r.moveTo(s+g+C+y*.5,u+d/2-y*.5),r.lineTo(s+g+C+y*.5,u+d/2+y*.5),r.lineWidth=2,r.strokeStyle=l.bgIconHeader,r.lineCap="round",r.stroke())}r.fillStyle=l.textMedium,r.fillText(t,h+s+l.cellHorizontalPadding+.5,u+d/2+Zn(r,l)),r.beginPath()}var gl=ke(Ie(),1);kt();var sN=gl.lazy(async()=>await Promise.resolve().then(()=>(ZE(),QE))),JE={getAccessibilityString:e=>e.data?.toString()??"",kind:te.Number,needsHover:!1,needsHoverPosition:!1,useLabel:!0,drawPrep:pi,draw:e=>Jn(e,e.cell.displayData,e.cell.contentAlign),measure:(e,t,n)=>e.measureText(t.displayData).width+n.cellHorizontalPadding*2,onDelete:e=>({...e,data:void 0}),provideEditor:()=>e=>{let{isHighlighted:t,onChange:n,value:r,validatedSelection:i}=e;return gl.createElement(gl.Suspense,{fallback:null},gl.createElement(sN,{highlight:t,disabled:r.readonly===!0,value:r.data,fixedDecimals:r.fixedDecimals,allowNegative:r.allowNegative,thousandSeparator:r.thousandSeparator,decimalSeparator:r.decimalSeparator,validatedSelection:i,onChange:o=>n({...r,data:Number.isNaN(o.floatValue??0)?0:o.floatValue})}))},onPaste:(e,t,n)=>{let r=typeof n.rawValue=="number"?n.rawValue:Number.parseFloat(typeof n.rawValue=="string"?n.rawValue:e);if(!(Number.isNaN(r)||t.data===r))return{...t,data:r,displayData:n.formattedString??t.displayData}}};kt();var e3={getAccessibilityString:()=>"",measure:()=>108,kind:te.Protected,needsHover:!1,needsHoverPosition:!1,draw:uN,onPaste:()=>{}};function uN(e){let{ctx:t,theme:n,rect:r}=e,{x:i,y:o,height:l}=r;t.beginPath();let a=2.5,s=i+n.cellHorizontalPadding+a,u=o+l/2,c=Math.cos($m(30))*a,d=Math.sin($m(30))*a;for(let f=0;f<12;f++)t.moveTo(s,u-a),t.lineTo(s,u+a),t.moveTo(s+c,u-d),t.lineTo(s-c,u+d),t.moveTo(s-c,u-d),t.lineTo(s+c,u+d),s+=8;t.lineWidth=1.1,t.lineCap="square",t.strokeStyle=n.textLight,t.stroke()}var t3=ke(Ie(),1);kt();var n3={getAccessibilityString:e=>e.data?.toString()??"",kind:te.RowID,needsHover:!1,needsHoverPosition:!1,drawPrep:(e,t)=>pi(e,t,e.theme.textLight),draw:e=>Jn(e,e.cell.data,e.cell.contentAlign),measure:(e,t,n)=>e.measureText(t.data).width+n.cellHorizontalPadding*2,provideEditor:()=>e=>{let{isHighlighted:t,onChange:n,value:r,validatedSelection:i}=e;return t3.default.createElement(go,{highlight:t,autoFocus:r.readonly!==!0,disabled:r.readonly!==!1,value:r.data,validatedSelection:i,onChange:o=>n({...r,data:o.target.value})})},onPaste:()=>{}};var r3=ke(Ie(),1);kt();fi();var i3={getAccessibilityString:e=>e.data?.toString()??"",kind:te.Text,needsHover:e=>e.hoverEffect===!0,needsHoverPosition:!1,drawPrep:pi,useLabel:!0,draw:e=>{let{cell:t,hoverAmount:n,hyperWrapping:r,ctx:i,rect:o,theme:l,overrideCursor:a}=e,{displayData:s,contentAlign:u,hoverEffect:c,allowWrapping:d}=t;if(c===!0&&n>0){i.textBaseline="alphabetic";let f=l.cellHorizontalPadding,h=l.cellVerticalPadding,p=Fi(s,i,l.baseFontFull,"alphabetic"),y=o.height-h,C=Math.min(y,p.actualBoundingBoxAscent*2.5);i.beginPath(),In(i,o.x+f/2,o.y+(o.height-C)/2+1,p.width+f*3,C-1,l.roundingRadius??4),i.globalAlpha=n,i.fillStyle=Ir(l.textDark,.1),i.fill(),i.globalAlpha=1,i.fillStyle=l.textDark,i.textBaseline="middle",a?.("text")}Jn(e,s,u,d,r)},measure:(e,t,n)=>{let r=t.displayData.split(`
 `,t.allowWrapping===!0?void 0:1),i=0;for(let o of r)i=Math.max(i,e.measureText(o).width);return i+2*n.cellHorizontalPadding},onDelete:e=>({...e,data:""}),provideEditor:e=>({disablePadding:e.allowWrapping===!0,editor:t=>{let{isHighlighted:n,onChange:r,value:i,validatedSelection:o}=t;return r3.createElement(go,{style:e.allowWrapping===!0?{padding:"3px 8.5px"}:void 0,highlight:n,autoFocus:i.readonly!==!0,disabled:i.readonly===!0,altNewline:!0,value:i.data,validatedSelection:o,onChange:l=>r({...i,data:l.target.value})})}}),onPaste:(e,t,n)=>e===t.data?void 0:{...t,data:e,displayData:n.formattedString??t.displayData}};var s3=ke(Ie(),1);var Qr=ke(Ie(),1);Yn();var o3=vt("div")({name:"UriOverlayEditorStyle",class:"gdg-u1rrojo",propsAsIs:!1});var cN=e=>{let{uri:t,onChange:n,forceEditMode:r,readonly:i,validatedSelection:o,preview:l}=e,[a,s]=Qr.useState(!i&&(t===""||r)),u=Qr.useCallback(()=>{s(!0)},[]);return a?Qr.createElement(go,{validatedSelection:o,highlight:!0,autoFocus:!0,value:t,onChange:n}):Qr.createElement(o3,null,Qr.createElement("a",{className:"gdg-link-area",href:t,target:"_blank",rel:"noopener noreferrer"},l),!i&&Qr.createElement("div",{className:"gdg-edit-icon",onClick:u},Qr.createElement(Ra,null)),Qr.createElement("textarea",{className:"gdg-input",autoFocus:!0}))},l3=cN;kt();function a3(e,t,n,r){let i=n.cellHorizontalPadding,o=t.height/2-e.actualBoundingBoxAscent/2,l=e.width,a=e.actualBoundingBoxAscent;return r==="right"?i=t.width-l-n.cellHorizontalPadding:r==="center"&&(i=t.width/2-l/2),{x:i,y:o,width:l,height:a}}var u3={getAccessibilityString:e=>e.data?.toString()??"",kind:te.Uri,needsHover:e=>e.hoverEffect===!0,needsHoverPosition:!0,useLabel:!0,drawPrep:pi,draw:e=>{let{cell:t,theme:n,overrideCursor:r,hoverX:i,hoverY:o,rect:l,ctx:a}=e,s=t.displayData??t.data,u=t.hoverEffect===!0;if(r!==void 0&&u&&i!==void 0&&o!==void 0){let c=Fi(s,a,n.baseFontFull),d=a3(c,l,n,t.contentAlign),{x:f,y:h,width:p,height:y}=d;if(i>=f-4&&i<=f-4+p+8&&o>=h-4&&o<=h-4+y+8){let C=Zn(a,n.baseFontFull);r("pointer");let g=5,m=h-C;a.beginPath(),a.moveTo(l.x+f,Math.floor(l.y+m+y+g)+.5),a.lineTo(l.x+f+p,Math.floor(l.y+m+y+g)+.5),a.strokeStyle=n.linkColor,a.stroke(),a.save(),a.fillStyle=e.cellFillColor,Jn({...e,rect:{...l,x:l.x-1}},s,t.contentAlign),Jn({...e,rect:{...l,x:l.x-2}},s,t.contentAlign),Jn({...e,rect:{...l,x:l.x+1}},s,t.contentAlign),Jn({...e,rect:{...l,x:l.x+2}},s,t.contentAlign),a.restore()}}a.fillStyle=u?n.linkColor:n.textDark,Jn(e,s,t.contentAlign)},onClick:e=>{let{cell:t,bounds:n,posX:r,posY:i,theme:o}=e,l=t.displayData??t.data;if(t.hoverEffect!==!0||t.onClickUri===void 0)return;let a=Ik(l,o.baseFontFull);if(a===void 0)return;let s=a3(a,n,o,t.contentAlign);wo({x:s.x-4,y:s.y-4,width:s.width+8,height:s.height+8},r,i)&&t.onClickUri(e)},measure:(e,t,n)=>e.measureText(t.displayData??t.data).width+n.cellHorizontalPadding*2,onDelete:e=>({...e,data:""}),provideEditor:e=>t=>{let{onChange:n,value:r,forceEditMode:i,validatedSelection:o}=t;return s3.createElement(l3,{forceEditMode:r.readonly!==!0&&(i||e.hoverEffect===!0&&e.onClickUri!==void 0),uri:r.data,preview:r.displayData??r.data,validatedSelection:o,readonly:r.readonly===!0,onChange:l=>n({...r,data:l.target.value})})},onPaste:(e,t,n)=>e===t.data?void 0:{...t,data:e,displayData:n.formattedString??t.displayData}};var c3=[DE,OE,yE,kE,bE,TE,ME,PE,JE,e3,n3,i3,u3];var h3=ke(f3(),1);var Zg=[],Jg=class extends yu{imageLoaded=()=>{};loadedLocations=[];cache={};setCallback(t){this.imageLoaded=t}sendLoaded=(0,h3.default)(()=>{this.imageLoaded(new mi(this.loadedLocations)),this.loadedLocations=[]},20);clearOutOfWindow=()=>{let t=Object.keys(this.cache);for(let n of t){let r=this.cache[n],i=!1;for(let o=0;o<r.cells.length;o++){let l=r.cells[o];if(this.isInWindow(l)){i=!0;break}}i?r.cells=r.cells.filter(this.isInWindow):(r.cancel(),delete this.cache[n])}};loadImage(t,n,r,i){let o=!1,l=Zg.pop()??new Image,a=!1,s={img:void 0,cells:[Hn(n,r)],url:t,cancel:()=>{a||(a=!0,Zg.length<12?Zg.unshift(l):o||(l.src=""))}},u=new Promise(c=>l.addEventListener("load",()=>c(null)));requestAnimationFrame(async()=>{try{l.src=t,await u,await l.decode();let c=this.cache[i];if(c!==void 0&&!a){c.img=l;for(let d of c.cells)this.loadedLocations.push(Ma(d));o=!0,this.sendLoaded()}}catch{s.cancel()}}),this.cache[i]=s}loadOrGetImage(t,n,r){let i=t,o=this.cache[i];if(o!==void 0){let l=Hn(n,r);return o.cells.includes(l)||o.cells.push(l),o.img}else this.loadImage(t,n,r,i)}},p3=Jg;var mN=(e,t)=>{let n=vl.useMemo(()=>({...NC,...e.headerIcons}),[e.headerIcons]),r=vl.useMemo(()=>e.imageWindowLoader??new p3,[e.imageWindowLoader]);return vl.createElement(gE,{...e,renderers:c3,headerIcons:n,ref:t,imageWindowLoader:r})},e0=vl.forwardRef(mN);var m3=`.gdg-mnuv029{word-break:break-word;-webkit-touch-callout:default;padding-top:6px}.gdg-mnuv029>*{margin:0}.gdg-mnuv029 *:last-child{margin-bottom:0}.gdg-mnuv029 p img{width:100%}.gdg-d19meir1{position:absolute;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-flex-direction:column;-ms-flex-direction:column;flex-direction:column;overflow:hidden;box-sizing:border-box;--overlay-top:var(--d19meir1-0);left:var(--d19meir1-1);top:var(--d19meir1-0);min-width:var(--d19meir1-2);min-height:var(--d19meir1-3);width:-webkit-max-content;width:-moz-max-content;width:max-content;max-width:400px;max-height:calc(100vh - var(--d19meir1-4));font-family:var(--gdg-font-family);font-size:var(--gdg-editor-font-size);text-align:start}@-webkit-keyframes glide_fade_in-gdg-d19meir1{0%{opacity:0%}to{opacity:100%}}@keyframes glide_fade_in-gdg-d19meir1{0%{opacity:0%}to{opacity:100%}}.gdg-d19meir1.gdg-style{border-radius:2px;background-color:var(--gdg-bg-cell);box-shadow:0 0 0 1px var(--gdg-accent-color),0 0 1px #3e415666,0 6px 12px #3e415626;-webkit-animation:glide_fade_in-gdg-d19meir1 60ms 1;animation:glide_fade_in-gdg-d19meir1 60ms 1}.gdg-d19meir1.gdg-pad{padding:var(--d19meir1-5) 8.5px 3px}.gdg-d19meir1 .gdg-clip-region{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-flex-direction:column;-ms-flex-direction:column;flex-direction:column;overflow-y:auto;overflow-x:hidden;border-radius:2px;-webkit-box-flex:1;-webkit-flex-grow:1;-ms-flex-positive:1;flex-grow:1}.gdg-d19meir1 .gdg-clip-region .gdg-growing-entry{height:100%}.gdg-d19meir1 .gdg-clip-region input.gdg-input{width:100%;border:none;border-width:0;outline:none}.gdg-d19meir1 .gdg-clip-region textarea.gdg-input{border:none;border-width:0;outline:none}.gdg-b1ygi5by{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-flex-wrap:wrap;-ms-flex-wrap:wrap;flex-wrap:wrap;margin-top:auto;margin-bottom:auto}.gdg-b1ygi5by .boe-bubble{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;border-radius:var(--gdg-rounding-radius,10px);padding:0 8px;height:20px;background-color:var(--gdg-bg-bubble);color:var(--gdg-text-dark);margin:2px}.gdg-b1ygi5by textarea{position:absolute;top:0;left:0;width:0px;height:0px;opacity:0}.gdg-n15fjm3e{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;margin:6px 0 3px;color:var(--gdg-text-dark)}.gdg-n15fjm3e>input{font-size:var(--gdg-editor-font-size);padding:0;font-family:var(--gdg-font-family);color:var(--gdg-text-dark);background-color:var(--gdg-bg-cell)}.gdg-d4zsq0x{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-flex-wrap:wrap;-ms-flex-wrap:wrap;flex-wrap:wrap}.gdg-d4zsq0x .doe-bubble{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;padding:0 8px;height:24px;background-color:var(--gdg-bg-cell);color:var(--gdg-text-dark);margin:2px;border-radius:var(--gdg-rounding-radius,6px);box-shadow:0 0 1px #3e415666,0 1px 3px #3e415666}.gdg-d4zsq0x .doe-bubble img{height:16px;object-fit:contain;margin-right:4px}.gdg-d4zsq0x textarea{position:absolute;top:0;left:0;width:0px;height:0px;opacity:0}.gdg-i2iowwq{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;height:100%}.gdg-i2iowwq .gdg-centering-container{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;height:100%}.gdg-i2iowwq .gdg-centering-container img,.gdg-i2iowwq .gdg-centering-container canvas{max-height:calc(100vh - var(--overlay-top) - 20px);object-fit:contain;-webkit-user-select:none;-moz-user-select:none;-ms-user-select:none;user-select:none}.gdg-i2iowwq .gdg-centering-container canvas{max-width:380px}.gdg-i2iowwq .gdg-edit-icon{position:absolute;top:12px;right:0;width:48px;height:48px;color:var(--gdg-accent-color);cursor:pointer;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center}.gdg-i2iowwq .gdg-edit-icon>*{width:24px;height:24px}.gdg-i2iowwq textarea{position:absolute;top:0;left:0;width:0px;height:0px;opacity:0}.gdg-u1rrojo{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-flex:1;-webkit-flex-grow:1;-ms-flex-positive:1;flex-grow:1;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;min-height:21px}.gdg-u1rrojo .gdg-link-area{-webkit-box-flex:1;-webkit-flex-grow:1;-ms-flex-positive:1;flex-grow:1;-webkit-flex-shrink:1;-ms-flex-negative:1;flex-shrink:1;cursor:pointer;margin-right:8px;overflow:hidden;text-overflow:ellipsis;white-space:nowrap;color:var(--gdg-link-color);-webkit-text-decoration:underline!important;text-decoration:underline!important}.gdg-u1rrojo .gdg-edit-icon{-webkit-flex-shrink:0;-ms-flex-negative:0;flex-shrink:0;width:32px;color:var(--gdg-accent-color);cursor:pointer;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center}.gdg-u1rrojo .gdg-edit-icon>*{width:24px;height:24px}.gdg-u1rrojo textarea{position:absolute;top:0;left:0;width:0px;height:0px;opacity:0}.gdg-m1pnx84e{min-width:var(--m1pnx84e-0);width:100%;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-align-items:flex-start;-webkit-box-align:flex-start;-ms-flex-align:flex-start;align-items:flex-start;-webkit-box-pack:justify;-webkit-justify-content:space-between;-ms-flex-pack:justify;justify-content:space-between;position:relative;color:var(--gdg-text-dark)}.gdg-m1pnx84e .gdg-g1y0xocz{-webkit-flex-shrink:1;-ms-flex-negative:1;flex-shrink:1;min-width:0}.gdg-m1pnx84e .gdg-spacer{-webkit-flex:1;-ms-flex:1;flex:1}.gdg-m1pnx84e .gdg-edit-icon{position:relative;cursor:pointer;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;color:var(--gdg-accent-color);padding:0;height:24px;width:24px;-webkit-flex-shrink:0;-ms-flex-negative:0;flex-shrink:0;-webkit-transition:all "0.125s ease";transition:all "0.125s ease";border-radius:6px}.gdg-m1pnx84e .gdg-edit-icon>*{width:16px;height:16px}.gdg-m1pnx84e .gdg-edit-hover:hover{background-color:var(--gdg-accent-light);-webkit-transition:background-color .15s;transition:background-color .15s}.gdg-m1pnx84e .gdg-checkmark-hover:hover{color:#fff;background-color:var(--gdg-accent-color)}.gdg-m1pnx84e .gdg-md-edit-textarea{position:relative;top:0;left:0;width:0px;height:0px;margin-top:25px;opacity:0;padding:0}.gdg-m1pnx84e .gdg-ml-6{margin-left:6px}.gdg-s1dgczr6 .dvn-scroller{overflow:var(--s1dgczr6-0);-webkit-transform:translate3d(0,0,0);-ms-transform:translate3d(0,0,0);transform:translateZ(0)}.gdg-s1dgczr6 .dvn-hidden{visibility:hidden}.gdg-s1dgczr6 .dvn-scroll-inner{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;pointer-events:none}.gdg-s1dgczr6 .dvn-scroll-inner>*{-webkit-flex-shrink:0;-ms-flex-negative:0;flex-shrink:0}.gdg-s1dgczr6 .dvn-scroll-inner .dvn-spacer{-webkit-box-flex:1;-webkit-flex-grow:1;-ms-flex-positive:1;flex-grow:1}.gdg-s1dgczr6 .dvn-scroll-inner .dvn-stack{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-flex-direction:column;-ms-flex-direction:column;flex-direction:column}.gdg-s1dgczr6 .dvn-underlay>*{position:absolute;left:0;top:0}.gdg-s1dgczr6 canvas{outline:none}.gdg-s1dgczr6 canvas *{height:0}.gdg-izpuzkl{position:absolute;inset:0;width:100%;height:100%;border-radius:0;resize:none;white-space:pre-wrap;min-width:100%;overflow:hidden;border:0;background-color:transparent;font-size:var(--gdg-editor-font-size);line-height:16px;font-family:var(--gdg-font-family);-webkit-text-fill-color:var(--gdg-text-dark);color:var(--gdg-text-dark);padding:0;margin:0}.gdg-izpuzkl::-webkit-input-placeholder{color:var(--gdg-text-light)}.gdg-izpuzkl::-moz-placeholder{color:var(--gdg-text-light)}.gdg-izpuzkl:-ms-input-placeholder{color:var(--gdg-text-light)}.gdg-izpuzkl::placeholder{color:var(--gdg-text-light)}.gdg-invalid .gdg-izpuzkl{-webkit-text-decoration:underline;text-decoration:underline;-webkit-text-decoration-color:#d60606;text-decoration-color:#d60606}.gdg-s69h75o{visibility:hidden;white-space:pre-wrap;word-wrap:break-word;width:-webkit-max-content;width:-moz-max-content;width:max-content;max-width:100%;min-width:100%;font-size:var(--gdg-editor-font-size);line-height:16px;font-family:var(--gdg-font-family);color:var(--gdg-text-dark);padding:0 0 2px;margin:0}.gdg-g1y0xocz{position:relative;margin-top:6px}.gdg-seveqep{position:absolute;top:4px;right:20px;background-color:var(--gdg-bg-cell);color:var(--gdg-text-dark);padding:8px;border:1px solid var(--gdg-border-color);border-radius:6px;font-size:var(--gdg-editor-font-size);-webkit-animation:gdg-search-fadein-gdg-seveqep .15s forwards;animation:gdg-search-fadein-gdg-seveqep .15s forwards}.gdg-seveqep.out{-webkit-animation:gdg-search-fadeout-gdg-seveqep .15s forwards;animation:gdg-search-fadeout-gdg-seveqep .15s forwards}.gdg-seveqep .gdg-search-bar-inner{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex}.gdg-seveqep .gdg-search-status{padding-top:4px;font-size:11px}.gdg-seveqep .gdg-search-progress{position:absolute;height:4px;left:0;bottom:0;background-color:var(--gdg-text-light)}.gdg-seveqep input{width:220px;color:var(--gdg-textDark);background-color:var(--gdg-bg-cell);border:none;border-width:0;outline:none}.gdg-seveqep button{width:24px;height:24px;padding:0;border:none;outline:none;background:none;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-box-pack:center;-webkit-justify-content:center;-ms-flex-pack:center;justify-content:center;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;cursor:pointer;color:var(--gdg-text-medium)}.gdg-seveqep button:hover{color:var(--gdg-text-dark)}.gdg-seveqep button .button-icon{width:16px;height:16px}.gdg-seveqep button:disabled{opacity:.4;pointer-events:none}@-webkit-keyframes gdg-search-fadeout-gdg-seveqep{0%{-webkit-transform:translateX(0);-ms-transform:translateX(0);transform:translate(0)}to{-webkit-transform:translateX(400px);-ms-transform:translateX(400px);transform:translate(400px)}}@keyframes gdg-search-fadeout-gdg-seveqep{0%{-webkit-transform:translateX(0);-ms-transform:translateX(0);transform:translate(0)}to{-webkit-transform:translateX(400px);-ms-transform:translateX(400px);transform:translate(400px)}}@-webkit-keyframes gdg-search-fadein-gdg-seveqep{0%{-webkit-transform:translateX(400px);-ms-transform:translateX(400px);transform:translate(400px)}to{-webkit-transform:translateX(0);-ms-transform:translateX(0);transform:translate(0)}}@keyframes gdg-search-fadein-gdg-seveqep{0%{-webkit-transform:translateX(400px);-ms-transform:translateX(400px);transform:translate(400px)}to{-webkit-transform:translateX(0);-ms-transform:translateX(0);transform:translate(0)}}.gdg-wmyidgi{position:relative;min-width:10px;min-height:10px;max-width:100%;max-height:100%;width:var(--wmyidgi-0);height:var(--wmyidgi-1);overflow:hidden;overflow:clip;direction:ltr}.gdg-wmyidgi>:first-child{position:absolute;left:0;top:0;width:100%;height:100%}.gdg-r17m35ur{-webkit-box-flex:1;-webkit-flex-grow:1;-ms-flex-positive:1;flex-grow:1;border:none;outline:none;background-color:var(--gdg-bg-header-has-focus);border-radius:9px;padding:0 8px;box-shadow:0 0 0 1px var(--gdg-border-color);color:var(--gdg-text-group-header);min-height:var(--r17m35ur-0);font:var(--gdg-header-font-style) var(--gdg-font-family)}.gdg-c1tqibwd{padding:0 8px;display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;-webkit-align-items:center;-webkit-box-align:center;-ms-flex-align:center;align-items:center;background-color:var(--gdg-bg-header)}
-`;var vN=/^-?\d+(?:\.\d+)?$/,yN=/^(\d{4})-(\d{2})-(\d{2})(?:[ T](\d{2}):(\d{2}):(\d{2}))?/;var wN=new Set(["true","t","1","yes","y"]),t0=e=>e.trim().toUpperCase(),g3=e=>t0(e).startsWith("DATE"),v3=e=>t0(e).startsWith("TIMESTAMP"),y3=e=>t0(e).startsWith("BOOL"),xN=e=>e.replace(/\B(?=(\d{3})+(?!\d))/g,","),w3=e=>{let t=String(e);if(!vN.test(t))return t;let n=t.startsWith("-"),r=n?t.slice(1):t,i=r.indexOf("."),o=i===-1?r:r.slice(0,i),l=i===-1?"":r.slice(i),a=xN(o)+l;return n?`-${a}`:a},n0=(e,t)=>{let n=String(e),r=yN.exec(n);if(!r)return n;let[,i,o,l,a,s,u]=r,c=`${i}-${o}-${l}`;return t&&a!==void 0?`${c} ${a}:${s}:${u}`:c},x3=e=>{if(e!=null)return typeof e=="boolean"?e:wN.has(String(e).trim().toLowerCase())};var SN=/^(DECIMAL|NUMERIC|NUMBER|INT|INTEGER|BIGINT|SMALLINT|TINYINT|DOUBLE|FLOAT|REAL)/i,kN=/^-?(?:\d+\.?\d*|\.\d+)(?:[eE][+-]?\d+)?$/,CN="(null)",RN=(e,t)=>t.find(n=>n.name===e)?.type??"",bN=(e,t)=>{let n=t.find(r=>r.name===e);return n!==void 0&&SN.test(n.type)},S3=e=>typeof e=="number"||typeof e=="string"&&kN.test(e),k3=(e,t,n,r)=>{let i=e[n],o=t[n];i==null&&(i=""),o==null&&(o="");let l=S3(i)?Number(i):NaN,a=S3(o)?Number(o):NaN;if(!isNaN(l)&&!isNaN(a))return r==="asc"?l-a:a-l;let s=String(i).toLowerCase(),u=String(o).toLowerCase();return r==="asc"?s<u?-1:s>u?1:0:u<s?-1:u>s?1:0},EN=e=>e==null?"":String(e),C3=(e,t,n,r)=>{if(e==null)return{kind:te.Text,data:"",displayData:CN,copyData:"",allowOverlay:!1,readonly:!0,themeOverride:{textDark:r}};let i=EN(e);if(bN(t,n)){let a=Number(e);if(!isNaN(a))return{kind:te.Number,data:a,displayData:w3(e),copyData:i,contentAlign:"left",allowOverlay:!1,readonly:!0}}let o=RN(t,n);if(y3(o)){let a=x3(e);if(a!==void 0)return{kind:te.Boolean,data:a,copyData:i,readonly:!0,allowOverlay:!1}}let l;return v3(o)?l=n0(e,!0):g3(o)?l=n0(e,!1):l=i,{kind:te.Text,data:i,displayData:l,copyData:i,allowOverlay:!0,readonly:!0}};var rr=(e,t)=>getComputedStyle(document.documentElement).getPropertyValue(e).trim()||t,r0=()=>{let e=rr("--vscode-editor-background","#1e1e1e"),t=rr("--vscode-foreground","#cccccc"),n=rr("--vscode-panel-border","rgba(128,128,128,0.35)"),r=rr("--vscode-editor-selectionBackground","rgba(38,79,120,0.4)"),i=rr("--vscode-focusBorder","#007fd4"),o=rr("--vscode-sideBarSectionHeader-background",e),l=rr("--vscode-editor-font-family","monospace"),a=rr("--vscode-editor-font-size","13px"),s=rr("--vscode-descriptionForeground","#999999");return{accentColor:i,accentFg:t,accentLight:r,textDark:t,textMedium:t,textLight:s,textBubble:t,bgIconHeader:o,fgIconHeader:t,textHeader:t,textHeaderSelected:t,bgCell:e,bgCellMedium:e,bgHeader:o,bgHeaderHasFocus:o,bgHeaderHovered:rr("--vscode-list-hoverBackground",o),borderColor:n,horizontalBorderColor:n,drilldownBorder:n,baseFontStyle:a,headerFontStyle:`600 ${a}`,fontFamily:l,editorFontSize:a}};var R3=e=>{let n=28+Math.min(Math.max(e,0),20)*28+28;return Math.min(n,560)};var Ba=ke(i0()),AN=150,M3=64,I3=400,_3="exasol-glide-style",l0="portal",LN=()=>{if(document.getElementById(_3))return;let e=document.createElement("style");e.id=_3,e.textContent=m3,document.head.appendChild(e)},Of=0,NN=()=>{if(Of+=1,document.getElementById(l0))return;let e=document.createElement("div");e.id=l0,e.style.position="fixed",e.style.left="0",e.style.top="0",e.style.zIndex="9999",document.body.appendChild(e)},o0=()=>{Of>0&&(Of-=1),!(Of>0)&&document.getElementById(l0)?.remove()},FN=e=>{let[t,n]=dt.useState(0);return dt.useEffect(()=>{let r=e.current;if(!r)return;let i=new ResizeObserver(o=>{let l=o[0].contentRect.width;n(a=>a===l?a:l)});return i.observe(r),()=>i.disconnect()},[e]),t},HN=()=>{let[e,t]=dt.useState(()=>r0());return dt.useEffect(()=>{let n=new MutationObserver(()=>t(r0()));return n.observe(document.body,{attributes:!0,attributeFilter:["class","style"]}),()=>n.disconnect()},[]),e},BN=({data:e})=>{let{columns:t,columnMetadata:n,rowCount:r,executionTime:i}=e,o=HN(),l=dt.useRef(null),a=FN(l),s=dt.useRef(null),[u,c]=dt.useState(null),[d,f]=dt.useState("asc"),[h,p]=dt.useState({}),[y,C]=dt.useState(()=>[...t]),[g,m]=dt.useState(void 0),[v,x]=dt.useState({current:void 0,rows:rt.empty(),columns:rt.empty()}),S=dt.useMemo(()=>{let N=new Set(t),W=y.filter(X=>N.has(X)),K=new Set(W);for(let X of t)K.has(X)||W.push(X);return W},[t,y]),R=dt.useMemo(()=>u?[...e.rows].sort((N,W)=>k3(N,W,u,d)):e.rows,[e.rows,u,d]),k=dt.useMemo(()=>S.map(N=>({title:N===u?`${N} ${d==="asc"?"\u25B2":"\u25BC"}`:N,id:N,width:h[N]??AN})),[S,h,u,d]),b=dt.useMemo(()=>rr("--vscode-descriptionForeground",o.textLight??"#999999"),[o]),M=dt.useCallback(N=>{let[W,K]=N,X=S[W],ge=R[K]?.[X];return C3(ge,X,n,b)},[R,S,n,b]),_=dt.useCallback(N=>{let W=[];for(let K=N.y;K<N.y+N.height;K++){let X=[];for(let ge=N.x;ge<N.x+N.width;ge++)X.push(M([ge,K]));W.push(X)}return W},[M]),L=dt.useCallback((N,W)=>{if(!N.id)return;let K=Math.min(I3,Math.max(M3,W));p(X=>({...X,[N.id]:K}))},[]),O=dt.useCallback((N,W)=>{if(N<0||N>=S.length||W<0||W>=S.length||N===W)return;let K=[...S],[X]=K.splice(N,1);K.splice(W,0,X),C(K)},[S]),D=dt.useCallback(N=>{let W=S[N];f(K=>u===W&&K==="asc"?"desc":"asc"),c(W)},[S,u]),A=dt.useCallback(N=>{let W=N.kind==="cell"?N.location[1]:void 0;m(K=>K===W?K:W)},[]),ne=dt.useMemo(()=>rr("--vscode-list-hoverBackground","rgba(128,128,128,0.12)"),[o]),Ce=dt.useCallback(N=>N===g?{bgCell:ne,bgCellMedium:ne}:void 0,[g,ne]),ce=R3(R.length),de=dt.useMemo(()=>rr("--vscode-descriptionForeground","#999999"),[o]),re=`${r.toLocaleString()} row(s) \u2014 ${i}ms`;return(0,Ba.jsxs)("div",{ref:l,className:"exasol-notebook-grid",style:{width:"100%"},children:[(0,Ba.jsx)("div",{className:"exasol-notebook-meta",style:{color:de,fontSize:"0.9em",padding:"2px 0 4px"},children:re}),a>0&&(0,Ba.jsx)(e0,{ref:s,columns:k,rows:R.length,getCellContent:M,getCellsForSelection:_,rowMarkers:"number",headerHeight:28,rowHeight:28,minColumnWidth:M3,maxColumnWidth:I3,width:a,height:ce,theme:o,smoothScrollX:!0,smoothScrollY:!0,gridSelection:v,onGridSelectionChange:x,rangeSelect:"multi-rect",onColumnResize:L,onColumnMoved:O,onHeaderClicked:D,onItemHovered:A,getRowThemeOverride:Ce,keybindings:{copy:!0}})]})},tU=e=>{LN();let t=new Map;return{renderOutputItem(n,r){let i=n.json(),o=t.get(n.id),l=o!==void 0&&o.element===r,a=l?o.root:(0,P3.createRoot)(r);o&&!l&&o.root.unmount(),l||NN(),o&&!l&&o0(),t.set(n.id,{element:r,root:a}),a.render((0,Ba.jsx)(BN,{data:i}))},disposeOutputItem(n){if(n===void 0){for(let{root:i}of t.values())i.unmount(),o0();t.clear();return}let r=t.get(n);r&&(r.root.unmount(),t.delete(n),o0())}}};export{tU as activate};
+`;var vN=/^-?\d+(?:\.\d+)?$/,yN=/^(\d{4})-(\d{2})-(\d{2})(?:[ T](\d{2}):(\d{2}):(\d{2}))?/;var wN=new Set(["true","t","1","yes","y"]),t0=e=>e.trim().toUpperCase(),g3=e=>t0(e).startsWith("DATE"),v3=e=>t0(e).startsWith("TIMESTAMP"),y3=e=>t0(e).startsWith("BOOL"),xN=e=>e.replace(/\B(?=(\d{3})+(?!\d))/g,","),w3=e=>{let t=String(e);if(!vN.test(t))return t;let n=t.startsWith("-"),r=n?t.slice(1):t,i=r.indexOf("."),o=i===-1?r:r.slice(0,i),l=i===-1?"":r.slice(i),a=xN(o)+l;return n?`-${a}`:a},n0=(e,t)=>{let n=String(e),r=yN.exec(n);if(!r)return n;let[,i,o,l,a,s,u]=r,c=`${i}-${o}-${l}`;return t&&a!==void 0?`${c} ${a}:${s}:${u}`:c},x3=e=>{if(e!=null)return typeof e=="boolean"?e:wN.has(String(e).trim().toLowerCase())};var SN=/^(DECIMAL|NUMERIC|NUMBER|INT|INTEGER|BIGINT|SMALLINT|TINYINT|DOUBLE|FLOAT|REAL)/i,kN=/^-?(?:\d+\.?\d*|\.\d+)(?:[eE][+-]?\d+)?$/,CN="(null)",RN=(e,t)=>t.find(n=>n.name===e)?.type??"";var bN=(e,t)=>{let n=t.find(r=>r.name===e);return n!==void 0&&SN.test(n.type)},S3=e=>typeof e=="number"||typeof e=="string"&&kN.test(e),k3=(e,t,n,r)=>{let i=e[n],o=t[n];i==null&&(i=""),o==null&&(o="");let l=S3(i)?Number(i):NaN,a=S3(o)?Number(o):NaN;if(!isNaN(l)&&!isNaN(a))return r==="asc"?l-a:a-l;let s=String(i).toLowerCase(),u=String(o).toLowerCase();return r==="asc"?s<u?-1:s>u?1:0:u<s?-1:u>s?1:0},EN=e=>e==null?"":String(e),C3=(e,t,n,r)=>{if(e==null)return{kind:te.Text,data:"",displayData:CN,copyData:"",allowOverlay:!1,readonly:!0,themeOverride:{textDark:r}};let i=EN(e);if(bN(t,n)){let a=Number(e);if(!isNaN(a))return{kind:te.Number,data:a,displayData:w3(e),copyData:i,contentAlign:"left",allowOverlay:!1,readonly:!0}}let o=RN(t,n);if(y3(o)){let a=x3(e);if(a!==void 0)return{kind:te.Boolean,data:a,copyData:i,readonly:!0,allowOverlay:!1}}let l;return v3(o)?l=n0(e,!0):g3(o)?l=n0(e,!1):l=i,{kind:te.Text,data:i,displayData:l,copyData:i,allowOverlay:!0,readonly:!0}};var rr=(e,t)=>getComputedStyle(document.documentElement).getPropertyValue(e).trim()||t,r0=()=>{let e=rr("--vscode-editor-background","#1e1e1e"),t=rr("--vscode-foreground","#cccccc"),n=rr("--vscode-panel-border","rgba(128,128,128,0.35)"),r=rr("--vscode-editor-selectionBackground","rgba(38,79,120,0.4)"),i=rr("--vscode-focusBorder","#007fd4"),o=rr("--vscode-sideBarSectionHeader-background",e),l=rr("--vscode-editor-font-family","monospace"),a=rr("--vscode-editor-font-size","13px"),s=rr("--vscode-descriptionForeground","#999999");return{accentColor:i,accentFg:t,accentLight:r,textDark:t,textMedium:t,textLight:s,textBubble:t,bgIconHeader:o,fgIconHeader:t,textHeader:t,textHeaderSelected:t,bgCell:e,bgCellMedium:e,bgHeader:o,bgHeaderHasFocus:o,bgHeaderHovered:rr("--vscode-list-hoverBackground",o),borderColor:n,horizontalBorderColor:n,drilldownBorder:n,baseFontStyle:a,headerFontStyle:`600 ${a}`,fontFamily:l,editorFontSize:a}};var R3=e=>{let n=28+Math.min(Math.max(e,0),20)*28+28;return Math.min(n,560)};var Ba=ke(i0()),AN=150,M3=64,I3=400,_3="exasol-glide-style",l0="portal",LN=()=>{if(document.getElementById(_3))return;let e=document.createElement("style");e.id=_3,e.textContent=m3,document.head.appendChild(e)},Of=0,NN=()=>{if(Of+=1,document.getElementById(l0))return;let e=document.createElement("div");e.id=l0,e.style.position="fixed",e.style.left="0",e.style.top="0",e.style.zIndex="9999",document.body.appendChild(e)},o0=()=>{Of>0&&(Of-=1),!(Of>0)&&document.getElementById(l0)?.remove()},FN=e=>{let[t,n]=dt.useState(0);return dt.useEffect(()=>{let r=e.current;if(!r)return;let i=new ResizeObserver(o=>{let l=o[0].contentRect.width;n(a=>a===l?a:l)});return i.observe(r),()=>i.disconnect()},[e]),t},HN=()=>{let[e,t]=dt.useState(()=>r0());return dt.useEffect(()=>{let n=new MutationObserver(()=>t(r0()));return n.observe(document.body,{attributes:!0,attributeFilter:["class","style"]}),()=>n.disconnect()},[]),e},BN=({data:e})=>{let{columns:t,columnMetadata:n,rowCount:r,executionTime:i}=e,o=HN(),l=dt.useRef(null),a=FN(l),s=dt.useRef(null),[u,c]=dt.useState(null),[d,f]=dt.useState("asc"),[h,p]=dt.useState({}),[y,C]=dt.useState(()=>[...t]),[g,m]=dt.useState(void 0),[v,x]=dt.useState({current:void 0,rows:rt.empty(),columns:rt.empty()}),S=dt.useMemo(()=>{let N=new Set(t),W=y.filter(X=>N.has(X)),K=new Set(W);for(let X of t)K.has(X)||W.push(X);return W},[t,y]),R=dt.useMemo(()=>u?[...e.rows].sort((N,W)=>k3(N,W,u,d)):e.rows,[e.rows,u,d]),k=dt.useMemo(()=>S.map(N=>({title:N===u?`${N} ${d==="asc"?"\u25B2":"\u25BC"}`:N,id:N,width:h[N]??AN})),[S,h,u,d]),b=dt.useMemo(()=>rr("--vscode-descriptionForeground",o.textLight??"#999999"),[o]),M=dt.useCallback(N=>{let[W,K]=N,X=S[W],ge=R[K]?.[X];return C3(ge,X,n,b)},[R,S,n,b]),_=dt.useCallback(N=>{let W=[];for(let K=N.y;K<N.y+N.height;K++){let X=[];for(let ge=N.x;ge<N.x+N.width;ge++)X.push(M([ge,K]));W.push(X)}return W},[M]),L=dt.useCallback((N,W)=>{if(!N.id)return;let K=Math.min(I3,Math.max(M3,W));p(X=>({...X,[N.id]:K}))},[]),O=dt.useCallback((N,W)=>{if(N<0||N>=S.length||W<0||W>=S.length||N===W)return;let K=[...S],[X]=K.splice(N,1);K.splice(W,0,X),C(K)},[S]),D=dt.useCallback(N=>{let W=S[N];f(K=>u===W&&K==="asc"?"desc":"asc"),c(W)},[S,u]),A=dt.useCallback(N=>{let W=N.kind==="cell"?N.location[1]:void 0;m(K=>K===W?K:W)},[]),ne=dt.useMemo(()=>rr("--vscode-list-hoverBackground","rgba(128,128,128,0.12)"),[o]),Ce=dt.useCallback(N=>N===g?{bgCell:ne,bgCellMedium:ne}:void 0,[g,ne]),ce=R3(R.length),de=dt.useMemo(()=>rr("--vscode-descriptionForeground","#999999"),[o]),re=`${r.toLocaleString()} row(s) \u2014 ${i}ms`;return(0,Ba.jsxs)("div",{ref:l,className:"exasol-notebook-grid",style:{width:"100%"},children:[(0,Ba.jsx)("div",{className:"exasol-notebook-meta",style:{color:de,fontSize:"0.9em",padding:"2px 0 4px"},children:re}),a>0&&(0,Ba.jsx)(e0,{ref:s,columns:k,rows:R.length,getCellContent:M,getCellsForSelection:_,rowMarkers:"number",headerHeight:28,rowHeight:28,minColumnWidth:M3,maxColumnWidth:I3,width:a,height:ce,theme:o,smoothScrollX:!0,smoothScrollY:!0,gridSelection:v,onGridSelectionChange:x,rangeSelect:"multi-rect",onColumnResize:L,onColumnMoved:O,onHeaderClicked:D,onItemHovered:A,getRowThemeOverride:Ce,keybindings:{copy:!0}})]})},tU=e=>{LN();let t=new Map;return{renderOutputItem(n,r){let i=n.json(),o=t.get(n.id),l=o!==void 0&&o.element===r,a=l?o.root:(0,P3.createRoot)(r);o&&!l&&o.root.unmount(),l||NN(),o&&!l&&o0(),t.set(n.id,{element:r,root:a}),a.render((0,Ba.jsx)(BN,{data:i}))},disposeOutputItem(n){if(n===void 0){for(let{root:i}of t.values())i.unmount(),o0();t.clear();return}let r=t.get(n);r&&(r.root.unmount(),t.delete(n),o0())}}};export{tU as activate};
 /*! Bundled license information:
 
 react/cjs/react.production.min.js:
